Move test dependencies in yac_mocks
