+++ test/module/irohad/consensus/yac/CMakeLists.txt
@@ -4,7 +4,7 @@ target_link_libraries(cluster_order_test
     yac
     )
 
-addtest(yac_simple_case_test yac_simple_case_test.cpp)
-target_link_libraries(yac_simple_case_test
+addtest(yac_cold_case_test yac_simple_cold_case_test.cpp)
+target_link_libraries(yac_cold_case_test
     yac
     )
