+++ test/module/irohad/consensus/yac/yac_mocks.hpp
@@ -21,8 +21,17 @@
 #include <gtest/gtest.h>
 #include "consensus/yac/yac.hpp"
 
-using namespace iroha::consensus::yac;
+using iroha::consensus::yac;
 using iroha::model::Peer;
+
+
+using iroha::model::Peer;
+
+Peer f_peer(std::string address) {
+  Peer peer;
+  peer.address = address;
+  return peer;
+};
 
 /**
  * Mock for yac crypto provider
@@ -112,6 +121,41 @@ class FakeNetwork : public YacNetwork {
   };
 
   std::shared_ptr<YacNetworkNotifications> notification;
+};
+
+class YacTest : public ::testing::Test {
+ public:
+  // ------|Netowrk|------
+  std::shared_ptr<FakeNetwork> network;
+  std::shared_ptr<CryptoProviderMock> crypto;
+  std::shared_ptr<FakeTimer> timer;
+  uint64_t delay = 100500;
+  std::shared_ptr<Yac> yac;
+
+  // ------|Round|------
+  std::vector<Peer> default_peers = {f_peer("1"),
+                                     f_peer("2"),
+                                     f_peer("3"),
+                                     f_peer("4"),
+                                     f_peer("5"),
+                                     f_peer("6"),
+                                     f_peer("7")
+  };
+
+  virtual void SetUp() override {
+    network = std::make_shared<FakeNetwork>();
+    crypto = std::make_shared<CryptoProviderMock>();
+    timer = std::make_shared<FakeTimer>();
+    yac = Yac::create(network,
+                      crypto,
+                      timer,
+                      delay);
+    network->subscribe(yac);
+  }
+
+  virtual void TearDown() override {
+    network->release();
+  }
 };
 
 #endif //IROHA_YAC_MOCKS_HPP
